Swiftlint fixes and code clean up
