+++ CodeEdit/Features/CEWorkspace/Models/CEWorkspaceFile.swift
@@ -47,9 +47,6 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
     /// If the item already is the top-level ``FileSystemClient/FileSystemClient/FileItem`` this returns `nil`.
     var parent: CEWorkspaceFile?
 
-    /// Returns the `id` in ``TabBarItemID`` enum form
-    var tabID: TabBarItemID { .codeEditor(id) }
-
     var fileDocument: CodeFileDocument?
 
     var fileIdentifier = UUID().uuidString
@@ -59,6 +56,9 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
 
     /// Returns the Git status of a file as ``GitType``
     var gitStatus: GitType?
+
+    /// Returns the `id` in ``TabBarItemID`` enum form
+    var tabID: TabBarItemID { .codeEditor(id) }
 
     /// Returns a boolean that is true if ``children`` is not `nil`
     var isFolder: Bool { url.hasDirectoryPath }
@@ -83,6 +83,18 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
             // item is a file
             return FileIcon.fileIcon(fileType: type)
         }
+    }
+
+    /// Return the file's UTType
+    var contentType: UTType? {
+        try? url.resourceValues(forKeys: [.contentTypeKey]).contentType
+    }
+
+    /// Returns a `Color` for a specific `fileType`
+    ///
+    /// If not specified otherwise this will return `Color.accentColor`
+    var iconColor: Color {
+        FileIcon.iconColor(fileType: type)
     }
 
     var debugFileHeirachy: String { childrenDescription(tabCount: 0) }
@@ -139,20 +151,6 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
         return true
     }
 
-    func childrenDescription(tabCount: Int) -> String {
-        var myDetails = "\(String(repeating: "|  ", count: max(tabCount - 1, 0)))\(tabCount != 0 ? "╰--" : "")"
-        myDetails += "\(url.path)"
-        if !self.isFolder { // if im a file, just return the url
-            return myDetails
-        } else { // if im a folder, return the url and its children's details
-            var childDetails = "\(myDetails)"
-            for child in children ?? [] {
-                childDetails += "\n\(child.childrenDescription(tabCount: tabCount + 1))"
-            }
-            return childDetails
-        }
-    }
-
     /// Returns a string describing a SFSymbol for folders
     ///
     /// If it is the top-level folder this will return `"square.dashed.inset.filled"`.
@@ -173,18 +171,6 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
         typeHidden ? url.deletingPathExtension().lastPathComponent : name
     }
 
-    /// Return the file's UTType
-    var contentType: UTType? {
-        try? url.resourceValues(forKeys: [.contentTypeKey]).contentType
-    }
-
-    /// Returns a `Color` for a specific `fileType`
-    ///
-    /// If not specified otherwise this will return `Color.accentColor`
-    var iconColor: Color {
-        FileIcon.iconColor(fileType: type)
-    }
-
     // MARK: Statics
     /// The default `FileManager` instance
     static let fileManger = FileManager.default
@@ -198,114 +184,6 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
     /// Allows the user to launch the file or folder as it would be in finder
     func openWithExternalEditor() {
         NSWorkspace.shared.open(url)
-    }
-
-    /// Flattens the children of ``self`` recursively with depth.
-    /// - Parameters:
-    ///   - depth: An int that indicates the how deep the tree files need to be flattened
-    ///   - ignoringFolders: A boolean on whether to ignore files that are Folders
-    /// - Returns: An array of flattened `CEWorkspaceFiles`
-    func flattenedChildren(withDepth depth: Int, ignoringFolders: Bool) -> [CEWorkspaceFile] {
-        guard depth > 0 else { return [] }
-        guard self.isFolder else { return [self] }
-        var childItems: [CEWorkspaceFile] = ignoringFolders ? [] : [self]
-        children?.forEach { child in
-            childItems.append(contentsOf: child.flattenedChildren(
-                withDepth: depth - 1,
-                ignoringFolders: ignoringFolders
-            ))
-        }
-        return childItems
-    }
-
-    /// Returns a list of `CEWorkspaceFiles` that are sibilings of ``self``.
-    /// The `height` parameter lets the function navigate up the folder hierarchy to
-    /// select a starting point from which it should start flettening the items.
-    /// - Parameters:
-    ///   - height: `Int` that tells where to start in the hierarchy
-    ///   - ignoringFolders: Wether the sibling folders should be flattened
-    /// - Returns: A list of `FileSystemItems`
-    func flattenedSiblings(withHeight height: Int, ignoringFolders: Bool) -> [CEWorkspaceFile] {
-        let topMostParent = self.getParent(withHeight: height)
-        return topMostParent.flattenedChildren(withDepth: height, ignoringFolders: ignoringFolders)
-    }
-
-    /// Using the current instance of `FileSystemItem` it will walk back up the Workspace file hiarchy
-    /// the amount of times specified with the `withHeight` parameter.
-    /// - Parameter height: The amount of times you want to up a folder.
-    /// - Returns: The found `FileSystemItem` object, This should always be a folder.
-    private func getParent(withHeight height: Int) -> CEWorkspaceFile {
-        var topmostParent = self
-        for _ in 0..<height {
-            guard let parent = topmostParent.parent else { break }
-            topmostParent = parent
-        }
-
-        return topmostParent
-    }
-
-    /// Recursive function that returns the number of children
-    /// that contain the `searchString` in their path or their subitems' paths.
-    /// Returns `0` if the item is not a folder.
-    /// - Parameters:
-    ///   - searchString: The string
-    ///   - ignoredStrings: The prefixes to ignore if they prefix file names
-    /// - Returns: The number of children that match the conditiions
-    func appearanceWithinChildrenOf(searchString: String, ignoredStrings: [String] = [".", "~"]) -> Int {
-        var count = 0
-        guard self.isFolder else { return 0 }
-        for child in self.children ?? [] {
-            var isIgnored: Bool = false
-            for ignoredString in ignoredStrings where child.name.hasPrefix(ignoredString) {
-                isIgnored = true // can use regex later
-            }
-
-            if isIgnored {
-                continue
-            }
-
-            guard !searchString.isEmpty else { count += 1; continue }
-            if child.isFolder {
-                count += child.appearanceWithinChildrenOf(searchString: searchString) > 0 ? 1 : 0
-            } else {
-                count += child.name.lowercased().contains(searchString.lowercased()) ? 1 : 0
-            }
-        }
-        return count
-    }
-
-    /// Function that returns an array of the children
-    /// that contain the `searchString` in their path or their subitems' paths.
-    /// Similar to `appearanceWithinChildrenOf(searchString: String)`
-    /// Returns `[]` if the item is not a folder.
-    /// - Parameter searchString: The string
-    /// - Parameter ignoredStrings: The prefixes to ignore if they prefix file names
-    /// - Returns: The children that match the conditiions
-    func childrenSatisfying(searchString: String, ignoredStrings: [String] = [".", "~"]) -> [CEWorkspaceFile] {
-        var satisfyingChildren: [CEWorkspaceFile] = []
-        guard self.isFolder else { return [] }
-        for child in self.children ?? [] {
-            var isIgnored: Bool = false
-            for ignoredString in ignoredStrings where child.name.hasPrefix(ignoredString) {
-                isIgnored = true // can use regex later
-            }
-
-            if isIgnored {
-                continue
-            }
-
-            guard !searchString.isEmpty else { satisfyingChildren.append(child); continue }
-            if child.isFolder {
-                if child.appearanceWithinChildrenOf(searchString: searchString) > 0 {
-                    satisfyingChildren.append(child)
-                }
-            } else {
-                if child.name.lowercased().contains(searchString.lowercased()) {
-                    satisfyingChildren.append(child)
-                }
-            }
-        }
-        return satisfyingChildren
     }
 
     /// This function allows creation of folders in the main directory or sub-folders
@@ -424,13 +302,11 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
                 previousName.replacingOccurrences(of: ".\(fileExtension)", with: "")
             fileUrl = fileUrl.deletingLastPathComponent().appendingPathComponent("\(fileName) copy\(fileExtension)")
         }
-//            Log.info("Duplicating file to \(fileUrl)")
 
         if CEWorkspaceFile.fileManger.fileExists(atPath: self.url.path) {
             do {
                 try CEWorkspaceFile.fileManger.copyItem(at: self.url, to: fileUrl)
             } catch {
-//                    Log.error("Error at \(self.url.path) to \(fileUrl.path)")
                 fatalError(error.localizedDescription)
             }
         }
@@ -442,7 +318,6 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
         createMissingParentDirectory(for: newLocation.deletingLastPathComponent())
 
         do {
-//                Log.info("Moving file \(self.url.debugDescription) to \(newLocation.debugDescription)")
             try CEWorkspaceFile.fileManger.moveItem(at: self.url, to: newLocation)
         } catch { fatalError(error.localizedDescription) }
 
@@ -454,7 +329,6 @@ final class CEWorkspaceFile: Codable, Comparable, Hashable, Identifiable, TabBar
             }
             // if the folder doesn't exist and the function was ordered to create it, create it.
             if createSelf && !CEWorkspaceFile.fileManger.fileExists(atPath: url.path) {
-//                    Log.info("Creating folder \(url.debugDescription)")
                 // Create the folder
                 do {
                     try CEWorkspaceFile.fileManger.createDirectory(
